Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.radial.inside1-expected.txt

2d.gradient.radial.inside1
Actual output:
Expected output:


FAIL Canvas test: 2d.gradient.radial.inside1 assert_equals: Green channel of the pixel at (98, 25) expected 255 but got 254
